Implement length-binned Dictionary class; improve performance
A number of significant changes were made.  First, the Dictionary class
should be the primary external API for dictionary.py - it implements
length-binning to ensure fast retrieval of words by length, at the cost
of some memory.  

Second, the Trie class has been updated - the tree traversal is now
iterative instead of recursve (just because it was prettier to look at
that way) and each Trie stores a master-list of all the words it
contains, so that if a pattern is not specified all words can be
returned quickly.  Also, Tries no longer have a conception of length of
words - that feature is provided only in the dictionary.  Tries are
designed to store arbitrary text and implement pattern-matching only.

Lastly, the Node class has been pared down to reduce memory usage
slightly.  Old features like parent and depth, which were necessary, are
no longer needed and were removed.