Test suite updated and fleshed out; performance testing improved
The testing module now uses the newly-implemented Dictionary class for
all tests.  It implements logging (currently set to level INFO) and has
a global boolean which switches off several tests if you want to test
only the performance of the Dictionary.  It also uses setUpClass for
initializing the Dictionary because I found the normal setUp to cause a
massive memory bloat.  It seemed like each time setUp was called, the
previous Dictionary was not being cleaned up, so memory ballooned
linearly with number of tests.  That doesn't make sense, but I don't
care, it happened and needed to stop.  Nothing modifies the class's
Dictionary during testing anyways, so each test is still independent.

The tests do not yet generate patterns or test for the non-existence of
non-words.  The first is critical, the second is less important but
useful.